requirements.txt
openai>=1.7.0
streamlit
pillow
pytesseract
python-dotenv
langdetect
